1, set recv and send len type from uint16 to uint32, let app layer to choose the real type 2, implement peer function in starnet, not tested 3, add some function to starnet.h, for peer issue
vet/app: me

+++ trunk/Server/server.cpp
@@ -111,7 +111,7 @@ void CALLBACK Server::OnDisconnect(ConnID connId)
 	pServer->m_pMainLoop->PushCommand(pCommand);
 }
 
-void CALLBACK Server::OnData(ConnID connId, uint16 iLen, char* pBuf)
+void CALLBACK Server::OnData(ConnID connId, uint32 iLen, char* pBuf)
 {
 	/*Connection* pConnection = (Connection*)connId;
 	Header* header = (Header*)pBuf;
